# spec/api-contract.yaml
openapi: 3.1.0
info:
  title: TTL-Aware BFF API
  version: 1.0.0
  summary: TTL-aware intelligent data-source routing Backend-for-Frontend.
  description: |
    Read-only aggregation tier over two heterogeneous sources:

      * **ODS** — Operational Data Source (gRPC, low latency, current state)
      * **EDS** — Execution Data Source (REST, higher latency, workflow / history / audit)

    Every request is classified, routed through an ordered 11-rule policy chain
    driven by configurable TTLs, source health, required fields and consistency
    requirements, mapped into a single canonical domain model, resolved by an
    explicit source-precedence policy, and returned inside a stable envelope
    carrying freshness and provenance metadata.

    Normative requirements: `spec/requirements.md`.
    Routing semantics: `spec/routing-policy.md`.
    Error semantics: `spec/error-model.md`.

    ## Response envelope

    Every 2xx body is `{ "data": ..., "meta": ... }` (REQ-API-004). `meta` always
    carries the routing decision, the stable routing rule id, the freshness
    verdict, per-field-group provenance, and structured warnings.

    ## Degradation

    `meta.degraded` means the data is old or came from a fallback source — both
    the routing-time rung (`health.primary_unavailable`) and the call-time rung
    (`fallback.primary_failed`) set it, each with a `SOURCE_UNAVAILABLE` warning.
    `meta.partial` means at least one source the request wanted did not answer.
...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /resources/{resourceId}/executions:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IncludeAudit'
      - $ref: '#/components/parameters/Consistency'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/TenantIdHeader'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/TraceParent'
    get:
      tags: [executions]
      operationId: listResourceExecutions
      summary: List a resource's execution history
      description: |
        Classification `execution_history`. Preferred source EXECUTION with
        `ttl: 0s` (**always live**), `cache_ttl: 0s`, `fallback: none`,
        `allow_stale: false`.

        `ttl: 0` means the freshness verdict is `STALE` by construction, so the
        fresh branch is unreachable and no cached or stale answer is ever served
        (REQ-TTL-010). Rule `fields.execution_only` fires because
        `executionHistory` can only be supplied by the EDS; no operational
        substitution is permitted, so EDS unavailability is a `503` rather than a
        degraded answer (REQ-EDGE-004).

        An existing resource with no executions returns `200` with
        `data.items: []` and `data.total: 0` — never `404` (REQ-EDGE-019). No
        warning is attached: an empty collection is an answer, and there is no
        `EMPTY_RESULT` code in the canonical warning vocabulary.

        Because `ttl: 0s`, `meta.freshness` reports `state: UNKNOWN` with
        `ttlSeconds: 0`. Execution data is a record of what happened rather than
        a cache of current state, so claiming `FRESH` would imply a guarantee the
        source does not make.
      x-requirement:
        - REQ-API-001
        - REQ-API-014
        - REQ-RT-001
        - REQ-TTL-010
        - REQ-EDGE-004
        - REQ-EDGE-019
      x-request-type: execution_history
      responses:
        '200':
          description: Execution history, newest first by `startedAt`.
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
            X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
            X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
            X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
            Cache-Control: { $ref: '#/components/headers/CacheControlOut' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionListEnvelope' }
              examples:
                page: { $ref: '#/components/examples/ExecutionHistoryPage' }
                empty: { $ref: '#/components/examples/ExecutionHistoryEmpty' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '502': { $ref: '#/components/responses/UpstreamInvalid' }
        '503': { $ref: '#/components/responses/NoSourceAvailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Required on every `/api/v1/*` route (REQ-SEC-001).

        Validation (REQ-SEC-002): signature against the JWKS at
        `security.jwt.jwks_url` (or HS256 from `security.jwt.hs256_secret_env`
        in development only), `iss`, `aud`, `exp`, `nbf`, `iat`, with
        `security.jwt.leeway` (default 30s) applied symmetrically. `alg` is
        allow-listed; `none` and algorithm-confusion attempts are rejected. All
        `security.jwt.required_claims` must be present.

        The tenant is resolved from the token claim (`tenant_id` by default) and
        is authoritative. When `X-Tenant-ID` is also supplied it must match the
        claim, otherwise `403 TENANT_MISMATCH` (REQ-SEC-004).

        Authorization is by role → permission mapping per endpoint
        (REQ-SEC-005); see `spec/security.md`.
  parameters:
    ResourceId:
      name: resourceId
      in: path
      required: true
      description: |
        Tenant-scoped resource identifier, the join key between ODS and EDS.
        Validated before anything else touches it: letters, digits and
        `- _ . :` only, at most 128 characters. Anything else is `400`, refused
        before a data source is contacted.
      schema:
        type: string
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
        minLength: 1
        maxLength: 128
      example: res-1042
...
    Limit:
      name: limit
      in: query
      required: false
      description: |
        Maximum executions to return.

        A non-integer or a value below 1 is rejected with
        `400 INVALID_REQUEST` (REQ-API-014). A value **above** the configured
        ceiling is clamped down to `sources.execution.max_history_items` (200)
        rather than rejected, because the ceiling is a protection for the source
        rather than a contract with the caller. Omitting it uses
        `sources.execution.history_page_size` (25).
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 25
      example: 25
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque pagination cursor returned as `data.nextCursor`. Base64url, at most 512 bytes.
      schema:
        type: string
        maxLength: 512
        pattern: '^[A-Za-z0-9_-]+={0,2}$'
      example: eyJvIjoxMDB9
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: |
        Client-supplied correlation id, echoed in the response header and in
        `meta.correlationId`. Generated as a UUIDv4 when absent or malformed
        (REQ-API-010). Propagated to both sources.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
      example: 2f9c1a7e-3b64-4c2a-9d15-7a0f4b8e1c33
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: |
        Optional tenant assertion. When present it MUST equal the JWT tenant
        claim, otherwise `403 TENANT_MISMATCH` (REQ-SEC-004).

        In a `local` or `test` environment running with
        `security.allow_insecure_no_auth: true`, which configuration validation
        permits nowhere else, this header *is* the tenant and no token is needed;
        it defaults to `local` when absent. That is how the compose stack is
        driven.
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
      example: acme
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Caller-side request identifier, logged but not interpreted.
      schema:
        type: string
        maxLength: 128
    TraceParent:
      name: traceparent
      in: header
      required: false
      description: W3C Trace Context parent. Honoured as the parent span and propagated to both sources (REQ-API-011).
      schema:
        type: string
        pattern: '^[0-9a-f]{2}-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$'
      example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    Consistency:
      name: consistency
      in: query
      required: false
      description: |
        Caller consistency override, honoured **only when it is stricter** than
        the level the endpoint is configured with: a caller may ask for more
        consistency than the endpoint promises, never for less (REQ-CLS-005).

        `strong` forbids stale-serve outright and makes rule
        `consistency.strong_requires_operational` fire, which reads the preferred
        source live with no TTL evaluation at all. It also disables the call-time
        fallback: the caller asked for one specific source's live answer and must
        not silently receive another's.

        It additionally **bypasses the cache read**. `Service.load` calls the
        loader directly rather than going through `GetOrLoad`, so `meta.cache.hit`
        is always `false`; the result is still written back via `Manager.Store`,
        because callers at weaker levels can legitimately use it (REQ-CACHE-006).

        There is no `Cache-Control: no-cache` handling on the request path; this
        query parameter is the only way to raise consistency.
      schema:
        type: string
        enum: [eventual, bounded, strong]
      example: strong
...
    IncludeAudit:
      name: includeAudit
      in: query
      required: false
      description: |
        Include audit records on execution reads. Requires the
        `execution.audit.read` permission: asking for audit data the caller may
        not see is `403`, not a silently narrowed response.
      schema:
        type: boolean
      example: false
  headers:
    XCorrelationID:
      description: Echo of the request correlation id, or the generated one (REQ-API-010).
      schema: { type: string }
    XBFFSource:
      description: |
        The routing target that answered, mirroring `meta.routingDecision`.
        Set on every 2xx by `response.Writer.OK`.
      schema:
        type: string
        enum: [OPERATIONAL, EXECUTION, BOTH, NONE]
    XBFFFreshness:
      description: Freshness verdict for the served data, mirroring `meta.freshness.state`.
      schema:
        type: string
        enum: [FRESH, STALE, UNKNOWN]
    XBFFDegraded:
      description: |
        Emitted as `true` only when `meta.degraded` is set, so an operator
        reading an access log can spot a degraded answer without a body parser.
        Absent otherwise. There is deliberately no routing-**rule** header: the
        rule id is a response-body field (`meta.routingRule`), not a header.
      schema:
        type: string
        enum: ['true']
    CacheControlOut:
      description: Always `no-store` for authenticated responses (REQ-SEC-014).
      schema: { type: string }
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        format: int32
        minimum: 0
    Allow:
      description: Methods permitted on this path.
      schema: { type: string }
  schemas:
...
    ExecutionListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/ExecutionList' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
  responses:
...
    InvalidRequest:
      description: Malformed path parameter, out-of-range `limit`, malformed cursor, or missing tenant.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            badLimit: { $ref: '#/components/examples/ErrInvalidRequest' }
            tenantMissing: { $ref: '#/components/examples/ErrTenantMissing' }
    Unauthenticated:
      description: Missing, malformed or invalid bearer token (REQ-SEC-002).
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            expired: { $ref: '#/components/examples/ErrUnauthenticated' }
    Forbidden:
      description: Authenticated but not permitted, or tenant assertion mismatch (REQ-SEC-004, REQ-SEC-005).
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            forbidden: { $ref: '#/components/examples/ErrForbidden' }
            tenantMismatch: { $ref: '#/components/examples/ErrTenantMismatch' }
...
    NotFound:
      description: |
        Every consulted source reports the resource absent (REQ-EDGE-019a).
        Negatively cached for `cache.negative_ttl`. An existing resource with an
        empty collection is `200`, not `404`.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            notFound: { $ref: '#/components/examples/ErrNotFound' }
    MethodNotAllowed:
      description: The path exists but the method is not GET (REQ-API-001).
      headers:
        Allow: { $ref: '#/components/headers/Allow' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: Token bucket exhausted for this tenant or globally (REQ-RES-008).
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            rateLimited: { $ref: '#/components/examples/ErrRateLimited' }
    Internal:
      description: Unexpected internal failure. Recovered panics land here (REQ-RES-012).
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UpstreamInvalid:
      description: |
        A source returned an undecodable or semantically invalid payload, or a
        major schema-version mismatch (REQ-EDGE-017, REQ-EDGE-020). Terminal —
        not retryable.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            invalidPayload: { $ref: '#/components/examples/ErrUpstreamInvalid' }
            schemaMismatch: { $ref: '#/components/examples/ErrSchemaMismatch' }
    NoSourceAvailable:
      description: |
        No source could produce data and the degradation ladder was exhausted
        (REQ-EDGE-005, REQ-RES-009). Retryable.
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            bothDown: { $ref: '#/components/examples/ErrNoSourceAvailable' }
            upstreamUnavailable: { $ref: '#/components/examples/ErrUpstreamUnavailable' }
...
  examples:
...
    ExecutionHistoryPage:
      summary: Execution history page (always live, ttl 0 / cache_ttl 0)
      description: |
        `execution_history` is configured `ttl: 0s` and `cache_ttl: 0s`, so this
        response is never cached and never stale-served. Freshness is `UNKNOWN`
        with `ttlSeconds: 0`: execution data is a record of what happened, not a
        cache of current state, and reporting `FRESH` would imply a guarantee the
        source does not make.
      value:
        data:
          resourceId: res-1042
          items:
            - executionId: exec-9c40de
              tenantId: acme
              resourceId: res-1042
              operation: RECONFIGURE
              status: RUNNING
              startedAt: '2026-09-03T11:58:12.004Z'
              updatedAt: '2026-09-03T12:00:41.771Z'
            - executionId: exec-8f21c4
              tenantId: acme
              resourceId: res-1042
              operation: RESIZE
              status: COMPLETED
              startedAt: '2026-09-02T09:14:02.100Z'
              completedAt: '2026-09-02T09:15:44.900Z'
              updatedAt: '2026-09-02T09:15:44.900Z'
              result: { outcome: SUCCEEDED, summary: 'RESIZE completed with state SUCCEEDED' }
          total: 7
          nextCursor: "2"
        meta:
          correlationId: ee55ff66-0077-4188-9299-334455667788
          routingDecision: EXECUTION
          routingRule: fields.execution_only
          sources: [EXECUTION]
          freshness:
            state: UNKNOWN
            ageSeconds: 0
            ttlSeconds: 0
            evaluatedAt: '2026-09-03T12:00:41.900Z'
            source: EXECUTION
          degraded: false
          partial: false
          cache: { hit: false, layer: NONE }
          warnings: []
          elapsedMs: 141
    ExecutionHistoryEmpty:
      summary: Existing resource with no executions — 200, not 404 (REQ-EDGE-019b)
      description: |
        An empty collection is an answer, so it is a `200` with an empty `items`
        array. There is no warning code for it: `EMPTY_RESULT` is not part of the
        canonical warning vocabulary.
      value:
        data:
          resourceId: res-1042
          items: []
          total: 0
        meta:
          correlationId: ff660077-1188-4299-a3aa-445566778899
          routingDecision: EXECUTION
          routingRule: fields.execution_only
          sources: [EXECUTION]
          freshness:
            state: UNKNOWN
            ageSeconds: 0
            ttlSeconds: 0
            evaluatedAt: '2026-09-03T12:00:42.100Z'
            source: EXECUTION
          degraded: false
          partial: false
          cache: { hit: false, layer: NONE }
          warnings: []
          elapsedMs: 133
...